LayoutTests/platform/mac-sierra/http/tests/inspector/network/resource-sizes-network-expected.txt

Test for Resource size values (transfer size, decoded size, header size, request and response) for resources served over the network.


== Running test suite: Resource.Size.Network
-- Running test case: Resource.Size.Network.text
PASS: statusCode should be 200.
PASS: compressed should be false.
PASS: responseSource should be Symbol(network).
PASS: estimatedNetworkEncodedSize should be exactly 2955 bytes.
PASS: estimatedTotalTransferSize should be >= (encoded body size + headers).
PASS: size should be exactly 2955 bytes.
FAIL: networkEncodedSize should be exactly 2955 bytes.
    Expected: 2955
    Actual: NaN
FAIL: networkTotalTransferSize should be >= (encoded body size + headers).
    Expected: greater than or equal to 2975
    Actual: NaN
FAIL: requestBodyTransferSize should be exactly 0 bytes.
    Expected: 0
    Actual: NaN
FAIL: responseBodyTransferSize should be exactly 2955 bytes.
    Expected: 2955
    Actual: NaN
FAIL: requestHeadersTransferSize should be non-empty.
    Expected: greater than 20
    Actual: NaN
FAIL: responseHeadersTransferSize should be non-empty.
    Expected: greater than 20
    Actual: NaN

-- Running test case: Resource.Size.Network.text.gzipped
PASS: statusCode should be 200.
PASS: compressed should be true.
PASS: responseSource should be Symbol(network).
PASS: estimatedNetworkEncodedSize should be exactly 1229 bytes.
PASS: estimatedTotalTransferSize should be >= (encoded body size + headers).
PASS: size should be exactly 2955 bytes.
FAIL: networkEncodedSize should be exactly 1229 bytes.
    Expected: 1229
    Actual: NaN
FAIL: networkTotalTransferSize should be >= (encoded body size + headers).
    Expected: greater than or equal to 1249
    Actual: NaN
FAIL: requestBodyTransferSize should be exactly 0 bytes.
    Expected: 0
    Actual: NaN
FAIL: responseBodyTransferSize should be exactly 1229 bytes.
    Expected: 1229
    Actual: NaN
FAIL: requestHeadersTransferSize should be non-empty.
    Expected: greater than 20
    Actual: NaN
FAIL: responseHeadersTransferSize should be non-empty.
    Expected: greater than 20
    Actual: NaN

-- Running test case: Resource.Size.Network.text.gzipped.no-content-length
PASS: statusCode should be 200.
PASS: compressed should be true.
PASS: responseSource should be Symbol(network).
FAIL: estimatedNetworkEncodedSize should be exactly 1229 bytes.
    Expected: 1229
    Actual: 2955
PASS: estimatedTotalTransferSize should be >= (encoded body size + headers).
PASS: size should be exactly 2955 bytes.
FAIL: networkEncodedSize should be exactly 1229 bytes.
    Expected: 1229
    Actual: NaN
FAIL: networkTotalTransferSize should be >= (encoded body size + headers).
    Expected: greater than or equal to 1249
    Actual: NaN
FAIL: requestBodyTransferSize should be exactly 0 bytes.
    Expected: 0
    Actual: NaN
FAIL: responseBodyTransferSize should be exactly 1229 bytes.
    Expected: 1229
    Actual: NaN
FAIL: requestHeadersTransferSize should be non-empty.
    Expected: greater than 20
    Actual: NaN
FAIL: responseHeadersTransferSize should be non-empty.
    Expected: greater than 20
    Actual: NaN

-- Running test case: Resource.Size.Network.image
PASS: statusCode should be 200.
PASS: compressed should be false.
PASS: responseSource should be Symbol(network).
PASS: estimatedNetworkEncodedSize should be exactly 12940 bytes.
PASS: estimatedTotalTransferSize should be >= (encoded body size + headers).
PASS: size should be exactly 12940 bytes.
FAIL: networkEncodedSize should be exactly 12940 bytes.
    Expected: 12940
    Actual: NaN
FAIL: networkTotalTransferSize should be >= (encoded body size + headers).
    Expected: greater than or equal to 12960
    Actual: NaN
FAIL: requestBodyTransferSize should be exactly 0 bytes.
    Expected: 0
    Actual: NaN
FAIL: responseBodyTransferSize should be exactly 12940 bytes.
    Expected: 12940
    Actual: NaN
FAIL: requestHeadersTransferSize should be non-empty.
    Expected: greater than 20
    Actual: NaN
FAIL: responseHeadersTransferSize should be non-empty.
    Expected: greater than 20
    Actual: NaN

-- Running test case: Resource.Size.Network.404
PASS: statusCode should be 404.
PASS: compressed should be false.
PASS: responseSource should be Symbol(network).
PASS: estimatedNetworkEncodedSize should be exactly 512 bytes.
PASS: estimatedTotalTransferSize should be >= (encoded body size + headers).
PASS: size should be exactly 512 bytes.
FAIL: networkEncodedSize should be exactly 512 bytes.
    Expected: 512
    Actual: NaN
FAIL: networkTotalTransferSize should be >= (encoded body size + headers).
    Expected: greater than or equal to 532
    Actual: NaN
FAIL: requestBodyTransferSize should be exactly 0 bytes.
    Expected: 0
    Actual: NaN
FAIL: responseBodyTransferSize should be exactly 512 bytes.
    Expected: 512
    Actual: NaN
FAIL: requestHeadersTransferSize should be non-empty.
    Expected: greater than 20
    Actual: NaN
FAIL: responseHeadersTransferSize should be non-empty.
    Expected: greater than 20
    Actual: NaN

-- Running test case: Resource.Size.Network.requestData
PASS: statusCode should be 200.
PASS: compressed should be false.
PASS: responseSource should be Symbol(network).
PASS: estimatedNetworkEncodedSize should be exactly 28 bytes.
PASS: estimatedTotalTransferSize should be >= (encoded body size + headers).
PASS: size should be exactly 28 bytes.
FAIL: networkEncodedSize should be exactly 28 bytes.
    Expected: 28
    Actual: NaN
FAIL: networkTotalTransferSize should be >= (encoded body size + headers).
    Expected: greater than or equal to 48
    Actual: NaN
FAIL: requestBodyTransferSize should be exactly 1027 bytes.
    Expected: 1027
    Actual: NaN
FAIL: responseBodyTransferSize should be exactly 28 bytes.
    Expected: 28
    Actual: NaN
FAIL: requestHeadersTransferSize should be non-empty.
    Expected: greater than 20
    Actual: NaN
FAIL: responseHeadersTransferSize should be non-empty.
    Expected: greater than 20
    Actual: NaN
FAIL: requestHeadersTransferSize should be > 1060 bytes.
    Expected: greater than 1060
    Actual: NaN
